LPS-158949 Move useCurrentKoroneikiAccount to comon directory
